test(server): split runs_test.go by concern (#424)
backend/internal/server/runs_test.go had grown to 1717 lines
covering legacy CRUD plus the workflow_spec (#411) and
issue_context (#415) surfaces interleaved by accretion. This
splits it into seven purpose-named files grouped by concern,
keeping the fakeRepo definition in a single helper file so the
test setup isn't duplicated.

  - runs_fake_test.go         fakeRepo + per-method seed helpers + newServer
  - runs_create_test.go       legacy TestCreateRun_* (happy path, idempotency, runner_kind)
  - runs_create_spec_test.go  TestCreateRun_WorkflowSpec_* + GitHub-fetch + spec fixtures
  - runs_create_issue_test.go TestCreateRun_IssueContext_*
  - runs_get_test.go          TestGetRun_* + TestRoundTrip_CreateThenGet
  - runs_list_test.go         TestListRuns_* (incl. runner_kind filter tests)
  - runs_cancel_test.go       TestCancelRun_*

runs_test.go shrinks to 28 lines, holding only the cross-cutting
TestErrorEnvelope_Shape. No test bodies changed; this is a
move-only refactor. Test count 53 → 53 (verified by counting
^func Test across both shapes). scripts/test passes;
golangci-lint clean.

Authored end-to-end via the agent-driven MCP local loop —
fishhawk_start_run → fishhawk_run_stage (plan) →
fishhawk_approve_plan → fishhawk_run_stage (implement) —
no terminal handoff.
