+++ backend/internal/server/runs_get_test.go
@@ -1,0 +1,138 @@
+package server
+
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"io"
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"testing"
+
+	"github.com/google/uuid"
+
+	"github.com/kuhlman-labs/fishhawk/backend/internal/run"
+)
+
+func TestGetRun_HappyPath(t *testing.T) {
+	repo := newFakeRepo()
+	s := newServer(t, repo)
+
+	// Pre-seed a run by calling the repo directly.
+	got, _ := repo.CreateRun(context.Background(), run.CreateRunParams{
+		Repo: "x/y", WorkflowID: "w", WorkflowSHA: "s",
+		TriggerSource: run.TriggerCLI,
+	})
+
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, fmt.Sprintf("/v0/runs/%s", got.ID), nil)
+	s.Handler().ServeHTTP(w, req)
+	if w.Code != http.StatusOK {
+		t.Fatalf("status = %d, want 200:\n%s", w.Code, w.Body.String())
+	}
+	var resp runResponse
+	_ = json.Unmarshal(w.Body.Bytes(), &resp)
+	if resp.ID != got.ID {
+		t.Errorf("ID = %s, want %s", resp.ID, got.ID)
+	}
+}
+
+func TestGetRun_NotFound(t *testing.T) {
+	s := newServer(t, newFakeRepo())
+	id := uuid.New()
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, fmt.Sprintf("/v0/runs/%s", id), nil)
+	s.Handler().ServeHTTP(w, req)
+	if w.Code != http.StatusNotFound {
+		t.Errorf("status = %d, want 404", w.Code)
+	}
+	if !strings.Contains(w.Body.String(), `"run_not_found"`) {
+		t.Errorf("body missing run_not_found code: %s", w.Body.String())
+	}
+}
+
+func TestGetRun_BadUUID(t *testing.T) {
+	s := newServer(t, newFakeRepo())
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, "/v0/runs/not-a-uuid", nil)
+	s.Handler().ServeHTTP(w, req)
+	if w.Code != http.StatusBadRequest {
+		t.Errorf("status = %d, want 400", w.Code)
+	}
+}
+
+func TestGetRun_RepoError(t *testing.T) {
+	repo := newFakeRepo()
+	repo.getErr = errors.New("connection lost")
+	s := newServer(t, repo)
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, fmt.Sprintf("/v0/runs/%s", uuid.New()), nil)
+	s.Handler().ServeHTTP(w, req)
+	if w.Code != http.StatusInternalServerError {
+		t.Errorf("status = %d, want 500", w.Code)
+	}
+}
+
+func TestGetRun_NilRepoConfigured(t *testing.T) {
+	s := New(Config{Addr: "127.0.0.1:0"})
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, fmt.Sprintf("/v0/runs/%s", uuid.New()), nil)
+	s.Handler().ServeHTTP(w, req)
+	if w.Code != http.StatusServiceUnavailable {
+		t.Errorf("status = %d, want 503", w.Code)
+	}
+}
+
+func TestRoundTrip_CreateThenGet(t *testing.T) {
+	s := newServer(t, newFakeRepo())
+
+	w, body := requestPath(t, s, http.MethodPost, "/v0/runs", map[string]any{
+		"repo":           "x/y",
+		"workflow_id":    "w",
+		"workflow_sha":   "abc",
+		"trigger_source": "ui",
+	})
+	if w.Code != http.StatusCreated {
+		t.Fatalf("create status = %d:\n%s", w.Code, body)
+	}
+	var created runResponse
+	if err := json.Unmarshal(body, &created); err != nil {
+		t.Fatal(err)
+	}
+
+	w2, body2 := requestPath(t, s, http.MethodGet, "/v0/runs/"+created.ID.String(), nil)
+	if w2.Code != http.StatusOK {
+		t.Fatalf("get status = %d:\n%s", w2.Code, body2)
+	}
+	var fetched runResponse
+	if err := json.Unmarshal(body2, &fetched); err != nil {
+		t.Fatal(err)
+	}
+	if fetched.ID != created.ID {
+		t.Errorf("ID round-trip mismatch: %s vs %s", fetched.ID, created.ID)
+	}
+}
+
+// requestPath is a tiny helper for round-tripping a raw body through
+// the server and asserting status + decoded JSON.
+func requestPath(t *testing.T, s *Server, method, path string, body any) (*httptest.ResponseRecorder, []byte) {
+	t.Helper()
+	var rdr io.Reader
+	if body != nil {
+		b, err := json.Marshal(body)
+		if err != nil {
+			t.Fatal(err)
+		}
+		rdr = bytes.NewReader(b)
+	}
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(method, path, rdr)
+	if body != nil {
+		req.Header.Set("Content-Type", "application/json")
+	}
+	s.Handler().ServeHTTP(w, req)
+	return w, w.Body.Bytes()
+}
